Implemented a "chunked" session proxy.
This "chunked" session proxy will allow a more responsive UI, for the tests done, the GTK UI, when one has a huge amount of torrents(>500).
What really happens is that the daemon is queried for chunks of torrent ids at a time, and not all torrents at the same time. Once finished, return the result.
Might probably need to also implement yielding the results, although I don't know if you can yield results from a deferred.