GTK UI loading speedup.
Limit the amount of data queried to the daemon to the torrents actually visible in the torrentview, those that are not visible, won't be queried until scrolled into a visible position. This speedup together with the chunked session proxy, makes the initial data requested from the daemon from ~35MB to about ~10MB for my > 800 torrents. And the time loading it has also dramatically dropped.