SessionProxy data restriction & Speedups.
The initial data queried by the SessionProxy is restricted to the visible columns by default on the GTK UI's torrent view, which itself only queries for it's visible columns status fields, no more, no less. No need to request data that it's not going to be needed. And if at a latter statge, that data is needed it's fetched like it previously was.