GTK UI Speedups. Fix TorrentView caching issues.
The caching stuff that was being done in the torrents view was removed, it was causing some KeyError issues on the SessionProxy and on several parts of the core.
Now, we're just requesting the minimum possible status fields, which, is still a major speed up. If this is not enough at a latter stage, I'll revisit this torrent view caching approach.